requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.30
python-dotenv==1.0.1
google-api-python-client==2.132.0
google-auth==2.31.0
google-auth-oauthlib==1.2.0
APScheduler==3.10.4
requests>=2.31.0
Jinja2>=3.1.4
